pyproject.toml
[project]
name = "parsidion"
version = "0.5.6"
requires-python = ">=3.13"
dependencies = []

[project.optional-dependencies]
search = [
    "fastembed>=0.6.0,<1.0",
    "sqlite-vec>=0.1.6,<1.0",
    "pillow>=12.2.0",
]
tools = [
    "parsidion[search]",
    "rich>=13.0",
]
eval = [
    "rich>=13.0",
    "pyyaml>=6.0",
    "fastembed>=0.6.0,<1.0",
    "sqlite-vec>=0.1.6,<1.0",
    "pillow>=12.2.0",
]

[project.scripts]
vault-search = "vault_search:main"
vault-new = "vault_new:main"
vault-stats = "vault_stats:main"
vault-review = "vault_review:main"
vault-export = "vault_export:main"
vault-merge = "vault_merge:main"

[build-system]
requires = [
    "setuptools>=70",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "vault_search",
    "vault_tui",
    "vault_common",
    "vault_new",
    "vault_stats",
    "vault_links",
    "vault_review",
    "vault_export",
    "vault_merge",
]

[tool.setuptools.package-dir]
"" = "skills/parsidion/scripts"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = "-v --cov=skills/parsidion/scripts --cov-report=term-missing"
timeout = 10

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "B",
    "BLE",
    "UP",
]
ignore = [
    "E501",
    "B008",
    "B904",
]

[tool.pyright]
extraPaths = [
    "skills/parsidion/scripts",
]
exclude = [
    "parsidion-mcp",
    ".worktrees",
    ".venv",
    "skills/parsidion/scripts/build_graph.py",
]

[tool.ty.environment]
extra-paths = [
    "skills/parsidion/scripts",
]

[tool.uv]
constraint-dependencies = [
    "requests>=2.33.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.3",
    "pytest-timeout>=2.4.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "pyright>=1.1.408",
    "rich>=14.3.4",
]
